+++ dremel3dpy/helpers/constants.py
@@ -6,8 +6,8 @@ import signal
 _LOGGER = logging.getLogger(__name__)
 
 MAJOR_VERSION = 1
-MINOR_VERSION = 0
-PATCH_VERSION = "1"
+MINOR_VERSION = 1
+PATCH_VERSION = "0"
 
 __version__ = f"{MAJOR_VERSION}.{MINOR_VERSION}.{PATCH_VERSION}"
 
@@ -95,7 +95,7 @@ FRAME_TEXT_FONT_SCALE = 0.6
 FRAME_TEXT_THICKNESS = 2
 FRAME_TEXT_COLOR = (0, 0, 0)
 
-DREMEL_MANUFACTURER = "Dremel Digilab"
+DREMEL_MANUFACTURER = "Dremel"
 
 MESSAGE = "message"
 ERROR_CODE = "error_code"